% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_functions.R
\name{assign_factor_level}
\alias{assign_factor_level}
\title{For a given data.frame and a data.frame of factors to reference
levels, assign all factors and ref levels in the main data.frame}
\usage{
assign_factor_level(df, factors)
}
\arguments{
\item{df}{data.frame with columns to make factors}

\item{factors}{A data.frame of factors to ref levels}
}
\value{
A data.frame with factors assigned
}
\description{
For a given data.frame and a data.frame of factors to reference
levels, assign all factors and ref levels in the main data.frame
}
