% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{coerce_formula}
\alias{coerce_formula}
\title{Try to convert the argument to a formula}
\usage{
coerce_formula(x)
}
\arguments{
\item{x}{A string, character vector, or formula}
}
\value{
A formula
}
\description{
Try to convert the argument to a formula
}
