% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_functions.R
\name{effect_c_stats}
\alias{effect_c_stats}
\title{Calculate the C statistic and a variety of associated stats
for a given survey model}
\usage{
effect_c_stats(survey_model, bubble_factor)
}
\arguments{
\item{survey_model}{The complete survey model}

\item{bubble_factor}{Multipying factor for bubble visualisation of model effects}
}
\value{
A dataframe of stats, along with the overall
model stats.
}
\description{
Calculate the C statistic and a variety of associated stats
for a given survey model
}
