% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_functions.R
\name{format_model_output}
\alias{format_model_output}
\title{Convert a Logit model into a dataframe in SDD ready formatted output.
Creates a data.frame with columns for:
\itemize{
\item questions,
\item question levels(i.e. sex = 1),
\item coefficients,
\item standard error,
\item Pr(>|t|) (the p-value),
\item confidence intervals,
\item odds ratios
}}
\usage{
format_model_output(model)
}
\arguments{
\item{model}{A fitted GLM}
}
\value{
A data.frame with columns:
}
\description{
Convert a Logit model into a dataframe in SDD ready formatted output.
Creates a data.frame with columns for:
\itemize{
\item questions,
\item question levels(i.e. sex = 1),
\item coefficients,
\item standard error,
\item Pr(>|t|) (the p-value),
\item confidence intervals,
\item odds ratios
}
}
