% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_functions.R
\name{sas_anova}
\alias{sas_anova}
\title{Custom anova stats to replicate SAS}
\usage{
sas_anova(survey_model)
}
\arguments{
\item{survey_model}{A svyglm model}
}
\value{
A data.frame of each effect and the ProbF
}
\description{
Custom anova stats to replicate SAS
}
