% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ergodic.R
\name{steady_state}
\alias{steady_state}
\alias{steady_state.default}
\alias{steady_state.sddr_markov}
\title{Ergodic (steady-state) distribution of a Markov chain}
\usage{
steady_state(x, ...)

\method{steady_state}{default}(x, ...)

\method{steady_state}{sddr_markov}(x, ...)
}
\arguments{
\item{x}{A row-stochastic matrix, or an object of class \code{sddr_markov}.}

\item{...}{Currently unused.}
}
\value{
A named numeric vector giving the stationary distribution.
}
\description{
Computes the long-run stationary distribution of a row-stochastic transition
matrix, i.e. the probability vector \eqn{\pi} satisfying
\eqn{\pi P = \pi} with \eqn{\sum_i \pi_i = 1}. It is obtained as the left
eigenvector of \code{P} associated with the eigenvalue closest to 1.
}
\examples{
P <- matrix(c(0.8, 0.2,
              0.3, 0.7), nrow = 2, byrow = TRUE)
steady_state(P)

}
