% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_functions.R
\name{survey_logit}
\alias{survey_logit}
\title{Create a survey design object and model a logistic regression
for a given formula. ALso calculates ANOVA and outputs.}
\usage{
survey_logit(data, formula, psu, strata, weight)
}
\arguments{
\item{data}{A dataset containing data needed for the model}

\item{formula}{A formula or string defining the model relationship}

\item{psu}{The ID/cluster column, as a string or formula}

\item{strata}{The strata column, as a string or formula}

\item{weight}{The weight column, as a string or formula}
}
\value{
A list of a model output from svyglm, and the ANOVA of that model
}
\description{
Used to ensure changes to the sdd survey logsitic regression
only need to be made once, and to allow strings to be passed
in instead of formulas for ease of use with rpy2 and command line.
}
