% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_functions.R
\name{survey_proportion}
\alias{survey_proportion}
\title{Calculate a weighted proportion of a variable, i.e. how often each value
of the variable occurs.}
\usage{
survey_proportion(data, variable, by, psu, strata, weights)
}
\arguments{
\item{data}{A dataset}

\item{variable}{A formula or string defining variable of interest}

\item{by}{A vector defining the subpopulations to calculate stats by, as
a string or a formula.}

\item{psu}{The ID/cluster column, as a string or formula}

\item{strata}{The strata column, as a string or formula}

\item{weights}{The weights column, as a string or formula}
}
\value{
data.frame
}
\description{
Calculate a weighted proportion of a variable, i.e. how often each value
of the variable occurs.
}
