% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_functions.R
\name{survey_ratio}
\alias{survey_ratio}
\title{Calculate a weighted ratio of one variable against another, as well as
standard errrors and confidence intervals}
\usage{
survey_ratio(data, variable, base, by, psu, strata, weights)
}
\arguments{
\item{data}{A dataset}

\item{variable}{A formula or string defining the numerator of the ratio}

\item{base}{A formula or string defining the denominator of the ratio}

\item{by}{A vector defining the subpopulations to calculate stats by, as
a string or a formula.}

\item{psu}{The ID/cluster column, as a string or formula}

\item{strata}{The strata column, as a string or formula}

\item{weights}{The weights column, as a string or formula}
}
\value{
data.frame
}
\description{
Calculate a weighted ratio of one variable against another, as well as
standard errrors and confidence intervals
}
