% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_functions.R
\name{survey_stats}
\alias{survey_stats}
\title{Calculate a set of weighted statistics for the variable.}
\usage{
survey_stats(data, variable, by, psu, strata, weights)
}
\arguments{
\item{data}{A dataset}

\item{variable}{A formula or string defining the variable of interest}

\item{by}{A vector defining the subpopulations to calculate stats by, as
a string or a formula.}

\item{psu}{The ID/cluster column, as a string or formula}

\item{strata}{The strata column, as a string or formula}

\item{weights}{The weight column, as a string or formula}
}
\value{
data.frame
}
\description{
Calculates columns: Mean, Median, std_err, lower_ci, upper_ci
}
