% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sddr-package.R
\docType{package}
\name{sddr-package}
\alias{sddr}
\alias{sddr-package}
\title{sddr: Spatial Distribution Dynamics}
\description{
\if{html}{\figure{logo.svg}{options: style='float: right' alt='logo' width='120'}}

A tidy toolkit for distribution dynamics: analysing how a cross-sectional distribution of values evolves over time and where it settles in the long run. Provides discrete-time, spatial, rank and local indicator of spatial association ('LISA') Markov transition estimation, ergodic analysis (steady-state, mean first passage and sojourn times), rank-mobility measures (Kendall's tau and the Theta statistic) and Markov mobility indices. Methods use long-format 'id'/'time'/'value' data rather than transition matrices and build on the distribution-dynamics literature (Quah (1993); Rey (2001) \doi{10.1111/j.1538-4632.2001.tb00444.x}). Results are validated for numerical parity against the reference 'giddy' library.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mqfarooqi1/sddr}
  \item Report bugs at \url{https://github.com/mqfarooqi1/sddr/issues}
}

}
\author{
\strong{Maintainer}: Muhammad Farooqi \email{mqfarooqi@gmail.com} (\href{https://orcid.org/0000-0003-4918-9791}{ORCID})

Authors:
\itemize{
  \item Muhammad Farooqi \email{mqfarooqi@gmail.com} (\href{https://orcid.org/0000-0003-4918-9791}{ORCID})
}

}
\keyword{internal}
